# OpenAPI-HL7-MIRWebservice.yaml
openapi: 3.0.0
info:
  title: Precision MIR API
  version: 1.0.0
servers:
  - url: https://precision.netgroup.uniroma2.it
    description: Precision server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /auth/token:
    post:
      summary: Request a Bearer Token
      tags:
        - authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "username"
                password:
                  type: string
                  example: "password"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkRlbW8gRGVtbyIsImlhdCI6MTUxNjIzOTAyMn0.nQswu-7zZz6dNRzkQ1ECcUVIll7dEyGZi9U1rHjSEqk"
                  tokenType:
                    type: string
                    example: "Bearer"
                  refreshToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkRlbW8gRGVtbyIsImlhdCI6MTUxNjIzOTAyMn0.nQswu-7zZz6dNRzkQ1ECcUVIll7dEyGZi9U1rHjSEqk"
                    nullable: true
                  expiresIn:
                    type: integer
                    example: 3600
                    description: Time in seconds before token expiration
                    nullable: true
  /patient:
    get:
      summary: Get the pending patient.
      tags:
        - patient
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                description: The patient data
                properties:
                  patient:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: "Mario Rossi"
                      birthdate:
                        type: string
                        example: "1980-01-01"
                  errorMessage:
                    type: string
                    example: null
                    nullable: true
        "400":
          description: Bad Request
  /results:
    post:
      summary: Send the patient results.
      tags:
        - results
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                results:
                  type: object
                  properties:
                    data:
                      type: object
                      description: Dati della spirometria
                      properties:
                        FVC:
                          type: number
                          example: 3.5
                        FEV1:
                          type: number
                          example: 2.5
                        PEF:
                          type: number
                          example: 5
                    pdf:
                      type: string
                      description: The base64 of the PDF generated by MIR Spiro
                      example: "abcdef=="
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  areResultsSaved:
                    type: boolean
                    example: true
                  error:
                    type: string
                    example: null
                    nullable: true
        "400":
          description: Bad Request